Moving the hooks to a new file
